--- apps/consultation-portal/components/Card/index.ts
@@ -1,1 +1,0 @@
-export * from './Card'
