+++ apps/consultation-portal/screens/Subscriptions/components/index.ts
@@ -4,3 +4,4 @@ export { default as TabsList } from './Tabs/tabsList'
 export { default as ChosenSubscriptions } from './ChosenSubscriptions/ChosenSubscriptions'
 export { default as SubscriptionChoices } from './SubscriptionChoices/SubscriptionChoices'
 export { default as SubscriptionCard } from './SubscriptionCard/SubscriptionCard'
+export { default as CardGridContainer } from './CardGridContainer/CardGridContainer'
